Remove user details connector / adapter and move methods to user connector / adapter
