+++ projects/core/src/occ/adapters/user/index.ts
@@ -1,7 +1,6 @@
 export * from './occ-user-address.adapter';
 export * from './occ-user.adapter';
 export * from './occ-user-consent.adapter';
-export * from './occ-user-details.adapter';
 export * from './occ-user-payment.adapter';
 export * from './occ-user-order.adapter';
 export * from './user-occ.module';
