+++ projects/core/src/occ/adapters/user/occ-user.adapter.ts
@@ -10,7 +10,7 @@ import {
 } from '../../utils/interceptor-util';
 import { ConverterService } from '../../../util/converter.service';
 import {
-  TITLE_NORMALIZER,
+  TITLE_NORMALIZER, USER_SERIALIZER,
   USER_SIGN_UP_SERIALIZER,
 } from '../../../user/connectors/user/converters';
 import { UserAdapter } from '../../../user/connectors/user/user.adapter';
@@ -36,6 +36,23 @@ export class OccUserAdapter implements UserAdapter {
     const endpoint = userId ? `${USER_ENDPOINT}${userId}` : USER_ENDPOINT;
     return this.occEndpoints.getEndpoint(endpoint);
   }
+
+  load(userId: string): Observable<User> {
+    const url = this.getUserEndpoint(userId);
+    return this.http.get<Occ.User>(url).pipe(
+      catchError((error: any) => throwError(error)),
+      this.converter.pipeable(USER_NORMALIZER)
+    );
+  }
+
+  update(userId: string, user: User): Observable<{}> {
+    const url = this.getUserEndpoint(userId);
+    user = this.converter.convert(user, USER_SERIALIZER);
+    return this.http
+      .patch(url, user)
+      .pipe(catchError(error => throwError(error)));
+  }
+
   register(user: UserSignUp): Observable<User> {
     const url: string = this.getUserEndpoint();
     let headers = new HttpHeaders({
