+++ projects/core/src/occ/adapters/user/user-occ.module.ts
@@ -4,8 +4,6 @@ import { HttpClientModule } from '@angular/common/http';
 
 import { UserOrderAdapter } from '../../../user/connectors/order/user-order.adapter';
 import { OccUserOrderAdapter } from './occ-user-order.adapter';
-import { UserDetailsAdapter } from '../../../user/connectors/details/user-details.adapter';
-import { OccUserDetailsAdapter } from './occ-user-details.adapter';
 import { UserAddressAdapter } from '../../../user/connectors/address/user-address.adapter';
 import { OccUserAddressAdapter } from './occ-user-address.adapter';
 import { UserAdapter } from '../../../user/connectors/user/user.adapter';
@@ -18,9 +16,8 @@ import { UserConsentAdapter } from '../../../user/connectors/consent/user-consen
 @NgModule({
   imports: [CommonModule, HttpClientModule],
   providers: [
-    { provide: UserDetailsAdapter, useClass: OccUserDetailsAdapter },
-    { provide: UserAddressAdapter, useClass: OccUserAddressAdapter },
     { provide: UserAdapter, useClass: OccUserAdapter },
+    { provide: UserAddressAdapter, useClass: OccUserAddressAdapter },
     { provide: UserConsentAdapter, useClass: OccUserConsentAdapter },
     {
       provide: UserPaymentAdapter,
