--- projects/core/src/user/connectors/details/index.ts
@@ -1,3 +1,0 @@
-export * from './user-details.connector';
-export * from './user-details.adapter';
-export * from './converters';
