--- projects/core/src/user/connectors/details/user-details.adapter.ts
@@ -1,7 +1,0 @@
-import { Observable } from 'rxjs';
-import { User } from '../../../model/misc.model';
-
-export abstract class UserDetailsAdapter {
-  abstract load(userId: string): Observable<User>;
-  abstract update(username: string, user: User): Observable<{}>;
-}
