+++ projects/core/src/user/connectors/user/converters.ts
@@ -1,6 +1,14 @@
 import { InjectionToken } from '@angular/core';
 import { Converter } from '../../../util/converter.service';
-import { Title, UserSignUp } from '../../../model/misc.model';
+import { Title, User, UserSignUp } from '../../../model/misc.model';
+
+export const USER_NORMALIZER = new InjectionToken<Converter<any, User>>(
+  'UserNormalizer'
+);
+
+export const USER_SERIALIZER = new InjectionToken<Converter<User, any>>(
+  'UserSerializer'
+);
 
 export const USER_SIGN_UP_SERIALIZER = new InjectionToken<
   Converter<UserSignUp, any>
