+++ projects/core/src/user/store/effects/reset-password.effect.spec.ts
@@ -10,8 +10,8 @@ import * as fromActions from './../actions';
 
 import { ResetPasswordEffects } from './reset-password.effect';
 import { GlobalMessageType, AddMessage } from '../../../global-message/index';
-import { UserConnector } from '../../connectors/user/user-account.connector';
-import { UserAdapter } from '../../connectors/user/user-account.adapter';
+import { UserConnector } from '../../connectors/user/user.connector';
+import { UserAdapter } from '../../connectors/user/user.adapter';
 
 describe('', () => {
   let service: UserConnector;
